chore(src): reorganize Pill elements
