+++ src/helix-ui/elements/hx-pill/_shadow.less
@@ -1,4 +1,4 @@
-@import "./HXElement";
+@import "../HXElement";
 @import (reference) "components/buttons";
 
 :host {
